requirements.txt
requests
autopep8
socketIO_client_nexus
web3
dotenv-python
pprint